Move login components back to top level folder
