+++ frontend/components/Splash.js
