+++ frontend/components/app.js
@@ -6,8 +6,8 @@ import {
   Text
 } from 'react-native';
 import { StackNavigator } from 'react-navigation';
-import LoginForm from './login/LoginForm.js';
-import Splash from './login/Splash.js';
+import LoginForm from './LoginForm.js';
+import Splash from './Splash.js';
 import Swiper from 'react-native-swiper';
 
 import CustomizeInterestList from './CustomizeInterestList/CustomizeInterestList.js';
@@ -15,26 +15,30 @@ import CustomizeInterestList from './CustomizeInterestList/CustomizeInterestList
 
 // Initialize three cards that the user can swipe between:
 // Splash Page, OAuth login, Normal Login
-const LegislateMe = React.createClass({
-  render: function(){
-    return (
-      <Swiper>
-        <View>
-          <Splash/>
-        </View>
-        <View>
-          <LoginForm/>
-        </View>
-        <View>
-          <CustomizeInterestList/>
-        </View>
-      </Swiper>
-    )
-  }
-});
-// const LegislateMe = StackNavigator({
-//   Splash: { screen: Splash },
-//   Login: {screen: LoginForm }
+
+
+// const LegislateMe = React.createClass({
+//   render: function(){
+//     return (
+//       <Swiper>
+//         <View>
+//           <Splash/>
+//         </View>
+//         <View>
+//           <LoginForm/>
+//         </View>
+//         <View>
+//           <CustomizeInterestList/>
+//         </View>
+//       </Swiper>
+//     )
+//   }
 // });
+
+
+const LegislateMe = StackNavigator({
+  Splash: { screen: Splash },
+});
+// Login: {screen: LoginForm }
 
 AppRegistry.registerComponent('legislate_me', () => LegislateMe);
